Cherry-pick 6969888: Implemented mGPU virtualization on a single GPU in D3D12, enabled via -vmgpu command line argument.
Virtualization works by creating multiple FD3D12Device-s, as if real SLI/LDA mode was used. All resources are created multiple times on the same physical device to simulate multiple LDA nodes.
Multiple 3D, copy and compute queues are created (one per device). Explicit transfers and synchronization must be performed to share resources between virtual devices, exactly like on a real mGPU machine.

This mode allows development and debugging of mGPU AFR and SFR features on any machine, using standard tools such as PIX, Nsight and RenderDoc that don't support real multi-node configurations.

Up to 4 GPUs can be virtualized this way (same number as maximum supported real LDA GPUs).

Example usage command line: MyProject -game -d3d12 -MaxGPUCount=2 -VMGPU -AFR

#rb Emil.Persson
#jira UE-76395

#ROBOMERGE-OWNER: ryan.vance
#ROBOMERGE-AUTHOR: juan.canada
#ROBOMERGE-SOURCE: CL 7053127 in //UE4/Release-4.23/... via CL 7053128
#ROBOMERGE-BOT: DEVVR (Main -> Dev-VR) (v367-6836689)

[CL 7146978 by juan canada in Dev-VR branch]

+++ Engine/Source/Runtime/D3D12RHI/Private/D3D12Allocation.cpp
@@ -1625,7 +1625,7 @@ FD3D12SegHeap* FD3D12SegList::CreateBackingHeap(
 	ID3D12Heap* D3DHeap;
 	D3D12_HEAP_DESC Desc = {};
 	Desc.SizeInBytes = HeapSize;
-	Desc.Properties = CD3DX12_HEAP_PROPERTIES(HeapType, Parent->GetGPUMask(), VisibleNodeMask);
+	Desc.Properties = CD3DX12_HEAP_PROPERTIES(HeapType, (uint32)Parent->GetGPUMask(), (uint32)VisibleNodeMask);
 	Desc.Flags = HeapFlags;
 
 	VERIFYD3D12RESULT(Parent->GetDevice()->CreateHeap(&Desc, IID_PPV_ARGS(&D3DHeap)));
